Preprocess input data only once
In the previous version, an example is preprocessed on the CPU each time `MeteorDataset.__getitem__` is called, leading to a bottleneck.
(This was intended to make it easy to implement data augmentation.)
This commit moves the preprocessing step from
`MeteorDataset.__getitem__` to `MeteorDataset.process_root` to speed up
data loading.
The preprocessing is also performed after converting a list of examples
to a single tensor to achieve more speed up.
It might be better to change the way of preprocessing data.